+++ src/translations/en/email-validator.php
@@ -1,0 +1,18 @@
+<?php
+/**
+ * Email Validator plugin for Craft CMS 3.x
+ *
+ * Email address validation for user registrations, custom forms and more.
+ *
+ * @link      https://github.com/lukeyouell
+ * @copyright Copyright (c) 2018 Luke Youell
+ */
+
+return [
+    'Invalid email format.' => 'Invalid email format.',
+    'MX records do not exist for this domain.' => 'MX records do not exist for this domain.',
+    'Role-based email addresses are not allowed.' => 'Role-based email addresses are not allowed.',
+    'Email addresses supplied by free providers are not allowed.' => 'Email addresses supplied by free providers are not allowed.',
+    'Disposable email addresses are not allowed.' => 'Disposable email addresses are not allowed.',
+    'Did you mean {suggestion}?' => 'Did you mean {suggestion}?',
+];